feat(filter/date-popover): update calendar item styles
