+++ src/utils/index.js
@@ -1,3 +1,4 @@
 export * from './function'
 export * from './string'
 export * from './tabs'
+export * from './time'
